変更　double => int

+++ Part10/UsersAnswer/bakeKitunE.cs
@@ -13,6 +13,6 @@ class Program
         DateTime afterTime = new DateTime(after[0], after[1], after[2], after[3], after[4], after[5]);
 
         TimeSpan time = afterTime - nowTime;
-        Console.WriteLine(time.TotalHours);
+        Console.WriteLine((int)time.TotalHours);
     }
 }
